refactor: move spec tests to individual files
